Refactoring: Move group and task views to common package
This change ends splitting `ansible_catalog.main.auth` package
into auth and common package.

* `ansible_catalog.main.auth` package contains only views related to the
authentication (e.g. current user profile and logout views).
* `ansible_catalog.main.common` contains views and models that are not
part of any `approval`, `catalog` or `inventory` subapps and provide
common API endpoints (e.g. roles, groups and tasks).
