+++ automation_services_catalog/main/auth/urls.py
@@ -1,12 +1,7 @@
 from django.urls import path
-from rest_framework.routers import SimpleRouter
 
 from automation_services_catalog.main.auth import views
 
-router = SimpleRouter()
-router.register("groups/sync", views.GroupSyncViewSet, basename="group-sync")
-router.register("groups", views.GroupViewSet, basename="group")
-router.register("tasks", views.TaskViewSet, basename="task")
 
 urlpatterns = [
     path(
@@ -14,5 +9,3 @@ urlpatterns = [
     ),
     path("logout/", views.SessionLogoutView.as_view(), name="logout"),
 ]
-
-urlpatterns += router.urls
