+++ automation_services_catalog/main/common/tests/functional/test_group_sync_task.py
@@ -9,8 +9,8 @@ from automation_services_catalog.common.auth.keycloak import (
     models as keycloak_models,
 )
 from automation_services_catalog.main.common import models
-from automation_services_catalog.main.auth import tasks
-from automation_services_catalog.main.auth.tests import factories
+from automation_services_catalog.main.common import tasks
+from automation_services_catalog.main.common.tests import factories
 
 
 @pytest.mark.django_db
